Fix AggregateException unwrap + code cleanup
