+++ AppDomainToolkit/AppDomainToolkit.csproj
@@ -62,6 +62,7 @@
     <Compile Include="IAssemblyTarget.cs" />
     <Compile Include="IAssemblyTargetLoader.cs" />
     <Compile Include="IDisposable.cs" />
+    <Compile Include="MarshalableTaskCompletionSource.cs" />
     <Compile Include="PathBasedAssemblyResolver.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.UnitTests.cs" />
